+++ crates/atuin/src/logs/otel.rs
@@ -1,149 +1,19 @@
 //! OpenTelemetry span export, gated behind the `profiling-traced` feature.
+//!
+//! Two symmetric `OtelCtx` implementations -- `enabled` (the real exporter) and `disabled` (a
+//! stub that errors if `ATUIN_OTEL` is set) -- keep the parent module free of `cfg`. Both expose
+//! `try_enable` and a `layer` returning `Box<dyn Layer<..>>`, so the signatures are identical.
+
 #[cfg(feature = "profiling-traced")]
-mod inner {
-    use std::env::VarError;
-    use std::ffi::OsString;
-    use std::str::FromStr;
-
-    use opentelemetry::trace::TracerProvider as _;
-    use opentelemetry_otlp::{ExporterBuildError, WithExportConfig as _};
-    use tracing_subscriber::Layer;
-    use tracing_subscriber::filter::LevelFilter;
-    use tracing_subscriber::registry::LookupSpan;
-    use url::Url;
-
-    #[derive(Debug, thiserror::Error)]
-    pub enum OtelCtxEnableError {
-        #[error("the given ATUIN_OTEL URL does not appear to be a utf-8 string")]
-        NonUtf8EnvVar(OsString),
-        #[error("the given ATUIN_OTEL failed to parse as URL: {0}")]
-        InvalidUrl(#[from] url::ParseError),
-        #[error("the given ATUIN_OTEL URL does not appear to be an HTTP(s) URL")]
-        NonHttpUrl,
-        #[error("failed to construct the exporter: {0}")]
-        ExporterBuild(#[from] ExporterBuildError),
-    }
-
-    pub struct OtelCtx {
-        provider: opentelemetry_sdk::trace::SdkTracerProvider,
-    }
-
-    impl OtelCtx {
-        /// Try to enable the opentelemetry logging context, if it was requested through the
-        /// `ATUIN_OTEL` environment variable.
-        pub fn try_enable(service_name: &'static str) -> Result<Option<Self>, OtelCtxEnableError> {
-            // TODO(markovejnovic): We should really have our own env-var parsing logic to avoid
-            // this annoying error handling here.
-            let otel_env: Option<String> = match std::env::var("ATUIN_OTEL") {
-                Ok(v) => Some(v),
-                Err(e) => match e {
-                    VarError::NotPresent => None,
-                    VarError::NotUnicode(e) => {
-                        return Err(OtelCtxEnableError::NonUtf8EnvVar(e));
-                    }
-                },
-            };
-
-            let otel_env: String = match otel_env {
-                Some(var) => var,
-                None => {
-                    return Ok(None);
-                }
-            };
-
-            // TODO(markovejnovic): A better env library could also handle this.
-            // TODO(markovejnovic): I want an HttpUrl type so bad...
-            let mut otel_url = Url::from_str(&otel_env)?;
-            if !(otel_url.scheme() == "http" || otel_url.scheme() == "https") {
-                return Err(OtelCtxEnableError::NonHttpUrl);
-            }
-
-            if !otel_url.path().ends_with("/v1/traces") {
-                otel_url
-                    .path_segments_mut()
-                    .expect("checked above that it's an http url")
-                    .extend(["v1", "traces"]);
-            }
-
-            let exporter = opentelemetry_otlp::SpanExporter::builder()
-                .with_http()
-                .with_protocol(opentelemetry_otlp::Protocol::HttpBinary)
-                .with_endpoint(otel_url.as_str())
-                .build()?;
-
-            let provider = opentelemetry_sdk::trace::SdkTracerProvider::builder()
-                .with_batch_exporter(exporter)
-                .with_resource(
-                    opentelemetry_sdk::Resource::builder().with_service_name(service_name).build(),
-                )
-                .build();
-
-            Ok(Some(Self { provider }))
-        }
-
-        /// The [`tracing_opentelemetry`] layer that exports spans through this context.
-        pub fn layer<S>(&self, service_name: &'static str) -> Box<dyn Layer<S> + Send + Sync>
-        where
-            S: tracing::Subscriber + for<'a> LookupSpan<'a> + Send + Sync,
-        {
-            Box::new(
-                tracing_opentelemetry::layer()
-                    .with_tracer(self.provider.tracer(service_name))
-                    .with_context_activation(false)
-                    .with_filter(LevelFilter::TRACE),
-            )
-        }
-    }
-
-    impl Drop for OtelCtx {
-        fn drop(&mut self) {
-            if let Err(err) = self.provider.force_flush() {
-                // Intentional eprintln! here since we cannot safely use error!
-                eprintln!("Unexpected error flushing OTEL spans: {err}");
-            }
-
-            if let Err(err) = self.provider.shutdown() {
-                // Intentional eprintln! here since we cannot safely use error!
-                eprintln!("Unexpected error shutting down the OTEL tracc provider: {err}");
-            }
-        }
-    }
-}
+mod enabled;
+#[cfg(feature = "profiling-traced")]
+pub(super) use enabled::OtelCtx;
+#[cfg(feature = "profiling-traced")]
+pub use enabled::OtelCtxEnableError;
 
 #[cfg(not(feature = "profiling-traced"))]
-mod inner {
-    use tracing_subscriber::Layer;
-    use tracing_subscriber::registry::LookupSpan;
-
-    #[derive(Debug, thiserror::Error)]
-    pub enum OtelCtxEnableError {
-        #[error(
-            "this build of atuin has no OpenTelemetry support: rebuild with the \
-             `profiling-traced` feature (e.g. `cargo build-traced`) to export `ATUIN_OTEL` traces"
-        )]
-        NotCompiled,
-    }
-
-    pub enum OtelCtx {}
-
-    impl OtelCtx {
-        /// Refuse to enable OpenTelemetry if it was requested, since it was compiled out.
-        pub fn try_enable(_service_name: &'static str) -> Result<Option<Self>, OtelCtxEnableError> {
-            if std::env::var_os("ATUIN_OTEL").is_some() {
-                return Err(OtelCtxEnableError::NotCompiled);
-            }
-            Ok(None)
-        }
-
-        #[allow(clippy::unused_self)]
-        pub fn layer<S>(&self, _service_name: &'static str) -> Box<dyn Layer<S> + Send + Sync>
-        where
-            S: tracing::Subscriber + for<'a> LookupSpan<'a> + Send + Sync,
-        {
-            unreachable!("OtelCtx is uninhabited without the profiling-traced feature")
-        }
-    }
-}
-
-pub(super) use inner::OtelCtx;
-pub use inner::OtelCtxEnableError;
+mod disabled;
+#[cfg(not(feature = "profiling-traced"))]
+pub(super) use disabled::OtelCtx;
+#[cfg(not(feature = "profiling-traced"))]
+pub use disabled::OtelCtxEnableError;
